+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lcap/builder
 
+## 1.1.2
+
+### Patch Changes
+
+- support sync in jsx, fix ref compile, add compiler optional params
+
+## 1.1.1
+
+### Patch Changes
+
+- ArrayExpression ObjectExpression transform to static
+
 ## 1.1.0
 
 ### Minor Changes