+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lcap/ui-cli
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @lcap/builder@1.1.2
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @lcap/builder@1.1.1
+
 ## 1.0.1
 
 ### Patch Changes